sentiment-classifier/requirements.txt
transformers==3.1.0
pytorch-lightning==0.9.0
pytorch-nlp==0.5.0
scikit-learn
pandas
nlpaug
faiss-cpu
sentence-transformers